# packages/cardano/connector-server-utxorpc/openapi.yaml
openapi: 3.1.0
info:
  title: Cardano Connector Server API
  version: 1.0.0
  description: |
    Mainnet-only HTTP facade over Dolos (UTxO RPC) and a bounded Koios
    transaction history capability.
servers:
  - url: http://localhost:8787
    description: Local development server
paths:
...
  /network:
    get:
      summary: Network
      operationId: getNetwork
      responses:
        "200":
          description: Fixed Mainnet identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkResponse"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /protocol-parameters:
    get:
      summary: Protocol parameters
      operationId: getProtocolParameters
      responses:
        "200":
          description: Live Bloxbean-compatible parameters from Dolos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtocolParametersResponse"
        default:
          $ref: "#/components/responses/UnexpectedError"
...
  /evaluate:
    post:
      summary: Evaluate a transaction without submitting it
      operationId: evaluateTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
      responses:
        "200":
          description: Node evaluation bound to the transaction body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationResponse"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /operations:
    post:
      summary: Create or reuse a durable L1 operation
      operationId: createOperation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOperationRequest"
      responses:
        "200":
          description: Operation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResponse"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /operations/{operation_id}:
    get:
      summary: Get durable L1 operation
      operationId: getOperation
      parameters:
        - name: operation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Operation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResponse"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /internal/channel-operations:
    post:
      summary: Create or reuse an adaptor-authorized channel operation
      operationId: createChannelOperation
      security:
        - channelOperator: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOperationRequest"
      responses:
        "200":
          description: Operation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResponse"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /internal/channel-operations/{operation_id}:
    get:
      summary: Get adaptor-authorized channel operation
      operationId: getChannelOperation
      security:
        - channelOperator: []
      parameters:
        - name: operation_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Operation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResponse"
        default:
          $ref: "#/components/responses/UnexpectedError"
components:
  securitySchemes:
    channelOperator:
      type: http
      scheme: bearer
...
  responses:
    UnexpectedError:
      description: Upstream failure or malformed request
      content:
        text/plain:
          schema:
            type: string
            example: unexpected error
  schemas:
...
    NetworkResponse:
      type: object
      additionalProperties: false
      required: [network]
      properties:
        network:
          type: string
          enum: [mainnet]
    ProtocolParametersResponse:
      type: object
      additionalProperties: false
      required: [era, epoch, slot, payload]
      properties:
        era:
          type: string
        epoch:
          type: integer
          minimum: 0
        slot:
          type: integer
          minimum: 0
        payload:
          $ref: "#/components/schemas/BloxbeanPayload"
    BloxbeanPayload:
      type: object
      additionalProperties: false
      required:
        - min_fee_a
        - min_fee_b
        - max_tx_size
        - key_deposit
        - pool_deposit
        - min_pool_cost
        - protocol_major_ver
        - protocol_minor_ver
        - coins_per_utxo_size
        - collateral_percent
        - max_collateral_inputs
      properties:
        min_fee_a:
          type: integer
          minimum: 0
        min_fee_b:
          type: integer
          minimum: 0
        max_tx_size:
          type: integer
          minimum: 0
        key_deposit:
          type: string
          maxLength: 20
          pattern: "^[0-9]+$"
        pool_deposit:
          type: string
          maxLength: 20
          pattern: "^[0-9]+$"
        min_pool_cost:
          type: string
          maxLength: 20
          pattern: "^[0-9]+$"
        protocol_major_ver:
          type: integer
          minimum: 0
        protocol_minor_ver:
          type: integer
          minimum: 0
        coins_per_utxo_size:
          type: string
          pattern: "^[0-9]+$"
          maxLength: 20
        collateral_percent:
          type: integer
          minimum: 0
        max_collateral_inputs:
          type: integer
          minimum: 0
...
    TxId:
      type: string
      pattern: "^[0-9a-f]{64}$"
...
    EvaluateRequest:
      type: object
      additionalProperties: false
      required: [transaction]
      properties:
        transaction:
          type: string
          pattern: "^(?:[0-9a-f]{2})+$"
          minLength: 2
          maxLength: 1048576
    EvaluationResponse:
      type: object
      additionalProperties: false
      required: [transaction_id, redeemers]
      properties:
        transaction_id:
          $ref: "#/components/schemas/TxId"
        redeemers:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/EvaluationRedeemer"
    EvaluationRedeemer:
      type: object
      additionalProperties: false
      required: [purpose, index, memory, steps]
      properties:
        purpose:
          type: string
          enum: [spend, mint, cert, reward, voting, proposing]
        index:
          type: integer
          minimum: 0
        memory:
          type: integer
          minimum: 0
        steps:
          type: integer
          minimum: 0
    CreateOperationRequest:
      type: object
      additionalProperties: false
      required: [operation_id, expected_transaction_id, transaction]
      properties:
        operation_id:
          type: string
          format: uuid
          pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
        expected_transaction_id:
          type: string
          pattern: "^[0-9a-f]{64}$"
        transaction:
          type: string
          pattern: "^(?:[0-9a-f]{2})+$"
          minLength: 2
          maxLength: 1048576
    OperationResponse:
      type: object
      additionalProperties: false
      required: [operation_id, expected_transaction_id, status, depth]
      properties:
        operation_id:
          type: string
          format: uuid
          pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
        expected_transaction_id:
          type: string
          pattern: "^[0-9a-f]{64}$"
        transaction_id:
          type: string
          pattern: "^[0-9a-f]{64}$"
        status:
          type: string
          enum: [pending, accepted, confirmed, settled, rejected]
        depth:
          type: integer
          minimum: 0